Refactor to PySide6 with logging
