+++ requirements.txt
@@ -1,0 +1,3 @@
+PySide6
+python-docx
+chardet
